backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
groq==0.9.0
sentence-transformers==3.0.1
hdbscan==0.8.38.post1
pgvector==0.3.2
statsmodels==0.14.4
pandas==2.2.3
numpy==2.2.4
python-dotenv==1.0.1
langgraph==0.2.73
langchain-core==0.3.66
pydantic==2.10.6
httpx==0.27.0
python-multipart==0.0.9
apscheduler==3.10.4
google-api-python-client==2.131.0
youtube-transcript-api==0.6.2
app-store-scraper==0.3.5
gnews==0.3.7
